Create Filter interface and BaseFilter abstract class
