+++ cameraview/src/main/java/com/otaliastudios/cameraview/filters/BrightnessFilter.java
@@ -5,7 +5,7 @@ import androidx.annotation.NonNull;
 /**
  * Adjusts the brightness of the preview.
  */
-public class BrightnessFilter extends Filter {
+public class BrightnessFilter extends BaseFilter {
     private float brightnessValue = 2.0f;
 
     /**
