+++ cameraview/src/main/java/com/otaliastudios/cameraview/filters/CrossProcessFilter.java
@@ -6,7 +6,7 @@ import androidx.annotation.NonNull;
  * Applies a cross process effect on preview, in which the red and green channels
  * are enhanced while the blue channel is restricted.
  */
-public class CrossProcessFilter extends Filter {
+public class CrossProcessFilter extends BaseFilter {
 
     /**
      * Initialize Effect
