+++ cameraview/src/main/java/com/otaliastudios/cameraview/filters/DuotoneFilter.java
@@ -7,7 +7,7 @@ import androidx.annotation.NonNull;
 /**
  * Representation of preview using only two color tones.
  */
-public class DuotoneFilter extends Filter {
+public class DuotoneFilter extends BaseFilter {
     // Default values
     private int mFirstColor = Color.MAGENTA;
     private int mSecondColor = Color.YELLOW;
