+++ cameraview/src/main/java/com/otaliastudios/cameraview/filters/NoFilter.java
@@ -2,7 +2,7 @@ package com.otaliastudios.cameraview.filters;
 
 import androidx.annotation.NonNull;
 
-public class NoFilter extends Filter {
+public class NoFilter extends BaseFilter {
 
     @NonNull
     @Override
